Remove helpers file/Achieve functionality from RecipeRepository class
